contract_analyzer/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pydantic==2.4.2
boto3==1.29.7
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.1.0
openai==1.3.5 